# backend/src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Campus Classroom Reservation System API
  version: 1.0.0
  description: >
    Campus classroom and seat reservation backend API.

    Authentication uses JWT Bearer tokens. `/auth/login` and `/auth/register`
    are public. Other endpoints require `Authorization: Bearer <token>`.

    Role model:
    - STUDENT: can self-register and reserve seats.
    - TEACHER: managed by administrators; can reserve seats and classrooms.
    - ADMIN: managed by administrators; manages classrooms, seats, and users.
      Admins do not participate in normal reservation flows in this MVP.

servers:
  - url: http://localhost:8080
    description: Local development server

tags:
  - name: Auth
    description: Login, student registration, and current user profile
  - name: Classroom
    description: Classroom and seat-layout query endpoints
  - name: Reservation
    description: Seat and classroom reservation endpoints
  - name: Admin
    description: Admin classroom and seat management
  - name: Seat
    description: Admin seat update endpoint

security:
  - bearerAuth: []

paths:
  /auth/login:
    post:
      tags: [Auth]
      summary: Login
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginDTO"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResultLoginVO"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /auth/register:
    post:
      tags: [Auth]
      summary: Register a student account
      description: >
        Public endpoint. This endpoint only creates STUDENT accounts.
        TEACHER and ADMIN accounts are assigned by administrators.
      operationId: registerStudent
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterDTO"
      responses:
        "200":
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResultVoid"
        "400":
          $ref: "#/components/responses/BadRequest"

  /auth/me:
    get:
      tags: [Auth]
      summary: Get current user profile
      operationId: getCurrentUser
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResultUserInfoVO"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /classrooms/available_list:
    get:
      tags: [Classroom]
      summary: List available classrooms
      operationId: listAvailableClassrooms
      parameters:
        - name: building
          in: query
          required: false
          schema:
            type: string
          description: Optional building filter
        - name: min_capacity
          in: query
          required: true
          schema:
            type: integer
            format: int32
            minimum: 1
          description: Minimum classroom capacity
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResultClassroomVOList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /classrooms/{id}:
    get:
      tags: [Classroom]
      summary: Get classroom by id
      operationId: getClassroomById
      parameters:
        - $ref: "#/components/parameters/ClassroomId"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResultClassroomVO"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  /classrooms/{id}/seats:
    get:
      tags: [Classroom]
      summary: Get classroom seat layout
      operationId: getClassroomSeatLayout
      parameters:
        - $ref: "#/components/parameters/ClassroomId"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResultClassroomSeatLayoutVO"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  /reservations/seats:
    post:
      tags: [Reservation]
      summary: Create seat reservation
      description: Requires STUDENT or TEACHER role.
      operationId: createSeatReservation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SeatReservationCreateDTO"
      responses:
        "200":
          description: Reservation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResultLong"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"

  /reservations/classrooms:
    post:
      tags: [Reservation]
      summary: Create classroom reservation
      description: Requires TEACHER role.
      operationId: createClassroomReservation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClassroomReservationCreateDTO"
      responses:
        "200":
          description: Reservation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResultLong"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"

  /reservations/{id}/cancel:
    delete:
      tags: [Reservation]
      summary: Cancel current user's reservation
      description: >
        Requires STUDENT or TEACHER role. Users can only cancel their own
        active, non-expired reservations.
      operationId: cancelReservation
      parameters:
        - $ref: "#/components/parameters/ReservationId"
      responses:
        "200":
          description: Reservation cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResultVoid"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"

  /reservations:
    get:
      tags: [Reservation]
      summary: List current user's active reservations
      description: Requires STUDENT or TEACHER role.
      operationId: listActiveReservations
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResultReservationVOList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

  /reservations/history:
    get:
      tags: [Reservation]
      summary: List current user's historical reservations
      description: Requires STUDENT or TEACHER role.
      operationId: listHistoryReservations
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResultReservationVOList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

  /admin/classrooms:
    post:
      tags: [Admin]
      summary: Create classroom
      description: Requires ADMIN role.
      operationId: adminCreateClassroom
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClassroomCreateDTO"
      responses:
        "200":
          description: Classroom created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResultClassroomVO"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"

  /admin/classrooms/list:
    get:
      tags: [Admin]
      summary: List classrooms for admin
      description: Requires ADMIN role.
      operationId: adminListClassrooms
      parameters:
        - name: building
          in: query
          required: true
          schema:
            type: string
        - name: min_capacity
          in: query
          required: true
          schema:
            type: integer
            format: int32
            minimum: 1
        - name: status
          in: query
          required: true
          schema:
            type: string
            enum: [ENABLED, DISABLED]
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResultClassroomVOList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

  /admin/classrooms/{id}:
    put:
      tags: [Admin]
      summary: Update classroom
      description: Requires ADMIN role.
      operationId: adminUpdateClassroom
      parameters:
        - $ref: "#/components/parameters/ClassroomId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClassroomUpdateDTO"
      responses:
        "200":
          description: Classroom updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResultClassroomVO"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  /admin/classrooms/{id}/seats/init:
    post:
      tags: [Admin]
      summary: Initialize seats for a classroom
      description: Requires ADMIN role.
      operationId: adminInitClassroomSeats
      parameters:
        - $ref: "#/components/parameters/ClassroomId"
      responses:
        "200":
          description: Seats initialized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResultVoid"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  /admin/classrooms/{id}/seats/status:
    put:
      tags: [Admin]
      summary: Batch update seat status in a classroom
      description: Requires ADMIN role.
      operationId: adminBatchUpdateSeatStatus
      parameters:
        - $ref: "#/components/parameters/ClassroomId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SeatUpdateDTO"
      responses:
        "200":
          description: Seat statuses updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResultVoid"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  /seats/{id}:
    put:
      tags: [Seat]
      summary: Update single seat
      description: Requires ADMIN role.
      operationId: adminUpdateSeat
      parameters:
        - $ref: "#/components/parameters/SeatId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SeatUpdateDTO"
      responses:
        "200":
          description: Seat updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResultSeatVO"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    ClassroomId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
      description: Classroom id
    SeatId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
      description: Seat id
    ReservationId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
      description: Reservation id

  responses:
    BadRequest:
      description: Invalid request parameter or business rule violation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResult"
    Unauthorized:
      description: Missing, invalid, or expired token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResult"
    Forbidden:
      description: Permission denied
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResult"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResult"
    Conflict:
      description: Reservation conflict or concurrent state change
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResult"

  schemas:
    ResultBase:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: success

    ErrorResult:
      allOf:
        - $ref: "#/components/schemas/ResultBase"
        - type: object
          properties:
            data:
              nullable: true
              example: null

    ResultVoid:
      allOf:
        - $ref: "#/components/schemas/ResultBase"
        - type: object
          properties:
            data:
              nullable: true
              example: null

    ResultLong:
      allOf:
        - $ref: "#/components/schemas/ResultBase"
        - type: object
          properties:
            data:
              type: integer
              format: int64
              example: 43

    ResultLoginVO:
      allOf:
        - $ref: "#/components/schemas/ResultBase"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/LoginVO"

    ResultUserInfoVO:
      allOf:
        - $ref: "#/components/schemas/ResultBase"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/UserInfoVO"

    ResultClassroomVO:
      allOf:
        - $ref: "#/components/schemas/ResultBase"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/ClassroomVO"

    ResultClassroomVOList:
      allOf:
        - $ref: "#/components/schemas/ResultBase"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/ClassroomVO"

    ResultClassroomSeatLayoutVO:
      allOf:
        - $ref: "#/components/schemas/ResultBase"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/ClassroomSeatLayoutVO"

    ResultSeatVO:
      allOf:
        - $ref: "#/components/schemas/ResultBase"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/SeatVO"

    ResultReservationVOList:
      allOf:
        - $ref: "#/components/schemas/ResultBase"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/ReservationVO"

    LoginDTO:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          example: student01
        password:
          type: string
          format: password
          example: Passw0rd!

    RegisterDTO:
      type: object
      required: [username, password, email]
      properties:
        username:
          type: string
          minLength: 6
          maxLength: 20
          example: student01
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 20
          example: Passw0rd!
        email:
          type: string
          format: email
          example: student01@example.com

    ClassroomCreateDTO:
      type: object
      required: [roomNumber, building]
      properties:
        roomNumber:
          type: string
          maxLength: 50
          example: A101
        building:
          type: string
          maxLength: 50
          example: Main Building
        seatRows:
          type: integer
          format: int32
          minimum: 1
          maximum: 100
          example: 8
        seatCols:
          type: integer
          format: int32
          minimum: 1
          maximum: 100
          example: 10
        status:
          type: string
          enum: [ENABLED, DISABLED]
          example: ENABLED
        remark:
          type: string
          maxLength: 255
          example: First floor

    ClassroomUpdateDTO:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [ENABLED, DISABLED]
          example: ENABLED
        remark:
          type: string
          maxLength: 255
          example: Under maintenance

    SeatUpdateDTO:
      type: object
      properties:
        status:
          type: string
          enum: [ENABLED, DISABLED]
          example: DISABLED
        remark:
          type: string
          maxLength: 255
          example: Broken chair

    SeatReservationCreateDTO:
      type: object
      required: [seat_id, start_time, end_time]
      properties:
        seat_id:
          type: integer
          format: int64
          example: 118
        start_time:
          type: string
          format: date-time
          example: "2026-04-22T19:00:00"
        end_time:
          type: string
          format: date-time
          example: "2026-04-22T21:00:00"
        reason:
          type: string
          maxLength: 255
          example: self-study

    ClassroomReservationCreateDTO:
      type: object
      required: [classroom_id, start_time, end_time]
      properties:
        classroom_id:
          type: integer
          format: int64
          example: 6
        start_time:
          type: string
          format: date-time
          example: "2026-04-22T19:00:00"
        end_time:
          type: string
          format: date-time
          example: "2026-04-22T21:00:00"
        reason:
          type: string
          maxLength: 255
          example: class meeting

    LoginVO:
      type: object
      properties:
        token:
          type: string
          example: eyJhbGciOiJIUzI1NiJ9.xxx.yyy
        userInfo:
          $ref: "#/components/schemas/UserInfoVO"

    UserInfoVO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10004
        username:
          type: string
          example: student01
        nickname:
          type: string
          example: student01
        email:
          type: string
          format: email
          example: student01@example.com
        role:
          type: string
          enum: [STUDENT, TEACHER, ADMIN]
          example: STUDENT

    ClassroomVO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 6
        roomNumber:
          type: string
          example: A101
        building:
          type: string
          example: Main Building
        capacity:
          type: integer
          format: int32
          example: 80
        seatRows:
          type: integer
          format: int32
          example: 8
        seatCols:
          type: integer
          format: int32
          example: 10
        status:
          type: string
          enum: [ENABLED, DISABLED]
          example: ENABLED
        remark:
          type: string
          example: First floor

    ClassroomSeatLayoutVO:
      type: object
      properties:
        classroomId:
          type: integer
          format: int64
          example: 6
        classroomBuilding:
          type: string
          example: Main Building
        classroomNumber:
          type: string
          example: A101
        seatRows:
          type: integer
          format: int32
          example: 8
        seatCols:
          type: integer
          format: int32
          example: 10
        seatVOS:
          type: array
          items:
            $ref: "#/components/schemas/SeatVO"

    SeatVO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 118
        classroomId:
          type: integer
          format: int64
          example: 6
        seatNumber:
          type: string
          example: "1-1"
        rowNumber:
          type: integer
          format: int32
          example: 1
        colNumber:
          type: integer
          format: int32
          example: 1
        status:
          type: string
          enum: [ENABLED, DISABLED]
          example: ENABLED
        remark:
          type: string
          example: Near window

    ReservationVO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 43
        resourceType:
          type: string
          enum: [SEAT, CLASSROOM]
          example: SEAT
        resourceId:
          type: integer
          format: int64
          example: 118
        resourceName:
          type: string
          example: Main Building A101 1-1
        reserveDate:
          type: string
          format: date
          example: "2026-04-22"
        startTime:
          type: string
          format: date-time
          example: "2026-04-22T19:00:00"
        endTime:
          type: string
          format: date-time
          example: "2026-04-22T21:00:00"
        reason:
          type: string
          example: self-study
        status:
          type: string
          enum: [ACTIVE, CANCELLED, EXPIRED]
          example: ACTIVE
        createTime:
          type: string
          format: date-time
          example: "2026-04-21T18:40:00"